+++ packages/jupyterlab-starters/src/widgets/fields/index.ts
@@ -1,0 +1,4 @@
+export { CodeMirrorField } from './codemirror';
+export { JSONObjectField } from './jsonobject';
+export { MarkdownField } from './markdown';
+export { XMLField } from './xml';