+++ packages/jupyterlab-starters/src/widgets/fields/markdown/index.tsx
@@ -1,0 +1,17 @@
+import * as React from 'react';
+
+import { CodeMirrorField } from '../codemirror';
+
+export function MarkdownField(props: any) {
+  const finalProps = {
+    ...props,
+    options: {
+      ...(props.options || {}),
+      cmOptions: {
+        ...(props.options?.cmOptions || {}),
+        mode: 'text/x-ipythongfm'
+      }
+    }
+  };
+  return <CodeMirrorField {...finalProps}></CodeMirrorField>;
+}