+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 - add notebook metadata authoring [#13][]
 - add Jupyter markdown to `title`, `description` and `ui:help` in schema forms [#17][]
 - all starters launch in right sidebar [#18][]
-- use CodeMirror for JSON editing [#21][]
+- enable CodeMirror for JSON, Markdown, etc. editing [#21][]
 
 ## `jupyter_starters 0.1.0a3`